+++ ToyBox/Classes/Infrastructure/Settings/GeneralSettings.cs
@@ -225,6 +225,7 @@ public class GeneralSettings : AbstractJsonSettings {
     public Dictionary<string, bool> OverrideEnableAiForCompanions = [];
     public bool EnableUnitOverrideMechanicalSize = false;
     public bool EnableUnitOverrideVisualSize = false;
+    public bool EnableUnitOverrideSkeleton = false;
     public bool EnableUnitModifyStats = false;
 
     // Dialog & NPCs